+++ R/agg_pipelines.R
@@ -9,44 +9,74 @@
 #' The exemplar table is assumed to be on the "exemplar_table" tab of the Excel file.
 #'
 #' @param countries A string vector of 3-letter country codes.
-#'                  Default is `NULL`, meaning all countries should be analyzed.
+#'                  Default is "all", meaning all available countries should be analyzed.
+#' @param years A numeric vector of years to be analyzed.
+#'              Default is "all", meaning all available years should be analyzed.
 #' @param psut_release The release we'll use from `psut_releases_folder`.
-#' @param psut_releases_folder The path to the `pins` archive of `PSUT` releases.
-#' @param exemplar_table_path The path to the exemplar table.
-#' @param world_agg_map The aggregation map to aggregate from continents to the world.
+#' @param aggregation_maps_path The path to the Excel file of aggregation maps.
+#' @param pipeline_caches_folder The path to a folder where .zip files of the targets pipeline are saved.
+#' @param pipeline_releases_folder The path to a folder where releases of output targets are pinned.
+#' @param release Boolean that tells whether to do a release of the results.
 #'
 #' @return A list of `tar_target`s to be executed in a workflow.
 #'
 #' @export
 get_pipeline <- function(countries = "all",
+                         years = "all",
                          psut_release,
-                         psut_releases_folder,
-                         exemplar_table_path,
-                         world_agg_map) {
+                         aggregation_maps_path,
+                         pipeline_caches_folder,
+                         pipeline_releases_folder,
+                         release = FALSE) {
 
   # Avoid notes when checking the package.
   keep_countries <- NULL
+  keep_years <- NULL
   PSUT <- NULL
   pinboard_folder <- NULL
   PSUT_release <- NULL
   PSUT_Re_continents <- NULL
   PSUT_Re_world <- NULL
-  world_aggregation_map <- NULL
   PSUT_Re_all <- NULL
   p_industry_prefixes <- NULL
   PSUT_Re_all_St_p <- NULL
   final_demand_sectors <- NULL
   PSUT_Re_all_St_fu <- NULL
   PSUT_Re_all_St_pfu <- NULL
+  PSUT_with_continent_col <- NULL
+  PSUT_Re_all_by_country <- NULL
+  PSUT_Re_all_St_pfu_by_country <- NULL
+  eta_Re_all_St_pfu <- NULL
+  aggregation_maps <- NULL
+
+  Continent <- NULL
+  Country <- NULL
+  map <- NULL
+
+  agg_eta_Re_all_St_pfu <- NULL
+  agg_Re_all_St_pfu <- NULL
+  pin_agg_csv <- NULL
+  pin_eta_csv <- NULL
+  pin_agg_eta_Re_all_St_pfu <- NULL
+  store_cache <- NULL
+  pipeline_caches_output_folder <- NULL
+
 
   # Create the pipeline
   list(
 
     # Identify the countries for this analysis.
-    # NULL means all countries.
+    # "all" means all countries.
     targets::tar_target(
       name = keep_countries,
       command = countries
+    ),
+
+    # Identify the countries for this analysis.
+    # "all" means all countries.
+    targets::tar_target(
+      name = keep_years,
+      command = years
     ),
 
     # Set the release that we'll use
@@ -55,10 +85,23 @@ get_pipeline <- function(countries = "all",
       command = psut_release
     ),
 
+    # Set the path to the aggregation maps file.
+    targets::tar_target_raw(
+      name = "aggregation_maps_path",
+      command = aggregation_maps_path
+    ),
+
+    # Set the folder for storing targets caches
+    targets::tar_target_raw(
+      name = "pipeline_caches_output_folder",
+      command = pipeline_caches_folder,
+      format = "file"
+    ),
+
     # Set the pinboard folder
     targets::tar_target_raw(
       "pinboard_folder",
-      psut_releases_folder,
+      pipeline_releases_folder,
       format = "file"
     ),
 
@@ -67,30 +110,28 @@ get_pipeline <- function(countries = "all",
       PSUT,
       pins::board_folder(pinboard_folder, versioned = TRUE) %>%
         pins::pin_read("psut", version = PSUT_release) %>%
-        filter_countries(keep_countries),
+        filter_countries_and_years(countries = keep_countries, years = keep_years),
       # Very important to assign storage and retrieval tasks to workers,
       # else the pipeline seemingly never finishes.
       storage = "worker",
       retrieval = "worker",
     ),
 
-    # Set the path to the exemplar_table,
-    # which we use for regional aggregations.
-    targets::tar_target_raw(
-      "exemplar_table_path",
-      exemplar_table_path
-    ),
 
+    ############################
+    # Prepare for aggregations #
+    ############################
 
-    #
-    # Regional aggregations
-    #
-
-    # Create the data frame to be used for continental aggregation
+    # Gather the aggregation maps.
     targets::tar_target(
-      continent_aggregation_map,
-      PFUAggDatabase::continent_aggregation_map(exemplar_table_path)
+      aggregation_maps,
+      load_aggregation_maps(path = aggregation_maps_path)
     ),
+
+
+    #########################
+    # Regional aggregations #
+    #########################
 
     # Create a continents data frame, grouped by continent,
     # so subsequent operations (region aggregation)
@@ -98,7 +139,7 @@ get_pipeline <- function(countries = "all",
     tarchetypes::tar_group_by(
       name = PSUT_with_continent_col,
       command = join_psut_continents(PSUT = PSUT,
-                                     continent_aggregation_map = continent_aggregation_map,
+                                     continent_aggregation_map = aggregation_maps$continent_aggregation,
                                      continent = "Continent"),
       # The columns to group by, as symbols.
       Continent,
@@ -111,26 +152,25 @@ get_pipeline <- function(countries = "all",
       PSUT_Re_continents,
       Recca::region_aggregates(PSUT_with_continent_col,
                                many_colname = IEATools::iea_cols$country,
-                               few_colname = "Continent"),
+                               few_colname = "Continent") %>%
+        # Eliminate the targets grouping on Continent
+        # so we can group by country later.
+        dplyr::mutate(
+          tar_group = NULL
+        ),
       pattern = map(PSUT_with_continent_col),
       storage = "worker",
       retrieval = "worker"
-    ),
-
-    # Create the world aggregation map,
-    # which is simply the incoming object.
-    targets::tar_target_raw(
-      "world_aggregation_map",
-      world_agg_map
     ),
 
     # Aggregate to world (WLD)
     targets::tar_target(
       PSUT_Re_world,
       Recca::region_aggregates(PSUT_Re_continents %>%
-                                 dplyr::mutate(
-                                   World = "WLD"
-                                 ),
+                                 dplyr::left_join(aggregation_maps$world_aggregation %>%
+                                                    matsbyname::agg_map_to_agg_table(many_colname = IEATools::iea_cols$country,
+                                                                                     few_colname = "World"),
+                                                  by = IEATools::iea_cols$country),
                                many_colname = IEATools::iea_cols$country, # Which actually holds continents
                                few_colname = "World")
     ),
@@ -142,20 +182,35 @@ get_pipeline <- function(countries = "all",
     ),
 
 
-    #
-    # PFU aggregations
-    #
+    ####################
+    # PFU aggregations #
+    ####################
+
+    # Set up a grouped-by-country data frame
+    # so all future calculations are parallelized across countries.
+    tarchetypes::tar_group_by(
+      name = PSUT_Re_all_by_country,
+      command = PSUT_Re_all,
+      # The columns to group by, as symbols.
+      Country,
+      storage = "worker",
+      retrieval = "worker"
+    ),
 
     # Establish prefixes for primary industries
     targets::tar_target(
       p_industry_prefixes,
-      IEATools::prim_agg_flows %>% unname() %>% unlist() %>% list()
+      IEATools::tpes_flows %>% unname() %>% unlist() %>% list()
     ),
 
     # Aggregate primary energy/exergy by total (total energy supply (TES)), product, and flow
     targets::tar_target(
       PSUT_Re_all_St_p,
-      calculate_primary_ex_data(PSUT_Re_all, p_industry_prefixes = p_industry_prefixes)
+      calculate_primary_ex_data(PSUT_Re_all_by_country,
+                                p_industry_prefixes = p_industry_prefixes),
+      pattern = map(PSUT_Re_all_by_country),
+      storage = "worker",
+      retrieval = "worker"
     ),
 
     # Establish final demand sectors
@@ -166,16 +221,119 @@ get_pipeline <- function(countries = "all",
 
     # Aggregate final and useful energy/exergy by total (total final consumption (TFC)), product, and sector
     targets::tar_target(
-      PSUT_Re_all_St_fu,
-      calculate_finaluseful_ex_data(.sutdata = PSUT_Re_all, fd_sectors = final_demand_sectors)
+      name = PSUT_Re_all_St_fu,
+      command = calculate_finaluseful_ex_data(PSUT_Re_all_by_country,
+                                              fd_sectors = final_demand_sectors),
+      pattern = map(PSUT_Re_all_by_country),
+      storage = "worker",
+      retrieval = "worker"
     ),
 
     # Bring the aggregations together in a single data frame
     targets::tar_target(
       PSUT_Re_all_St_pfu,
       dplyr::bind_rows(PSUT_Re_all_St_p, PSUT_Re_all_St_fu)
-    )
+    ),
 
+
+    ################
+    # Efficiencies #
+    ################
+
+    tarchetypes::tar_group_by(
+      name = PSUT_Re_all_St_pfu_by_country,
+      command = PSUT_Re_all_St_pfu %>%
+        dplyr::mutate(
+          tar_group = NULL
+        ),
+      # The columns to group by, as symbols.
+      Country,
+      storage = "worker",
+      retrieval = "worker"
+    ),
+
+    targets::tar_target(
+      name = agg_eta_Re_all_St_pfu,
+      command = calc_agg_etas(PSUT_Re_all_St_pfu_by_country),
+      pattern = map(PSUT_Re_all_St_pfu_by_country),
+      storage = "worker",
+      retrieval = "worker"
+    ),
+
+    # Split the aggregations and efficiencies apart
+    # to enable easier saving of separate .csv files later.
+    targets::tar_target(
+      name = agg_Re_all_St_pfu,
+      command = agg_eta_Re_all_St_pfu %>%
+        dplyr::mutate(
+          "{PFUAggDatabase::efficiency_cols$eta_pf}" := NULL,
+          "{PFUAggDatabase::efficiency_cols$eta_fu}" := NULL,
+          "{PFUAggDatabase::efficiency_cols$eta_pu}" := NULL,
+        ),
+      storage = "worker",
+      retrieval = "worker"
+    ),
+
+    targets::tar_target(
+      name = eta_Re_all_St_pfu,
+      command = agg_eta_Re_all_St_pfu %>%
+        dplyr::mutate(
+          "{IEATools::all_stages$primary}" := NULL,
+          "{IEATools::all_stages$final}" := NULL,
+          "{IEATools::all_stages$useful}" := NULL,
+        ),
+      storage = "worker",
+      retrieval = "worker"
+    ),
+
+
+    ################
+    # Save results #
+    ################
+
+    # Pin the aggregates and efficiencies as an .rds file
+    targets::tar_target(
+      pin_agg_eta_Re_all_St_pfu,
+      PFUWorkflow::release_target(pipeline_releases_folder = pinboard_folder,
+                                  targ = agg_eta_Re_all_St_pfu,
+                                  targ_name = "agg_eta_Re_all_St_pfu",
+                                  release = release)
+    ),
+
+    # Pin aggregates as a wide-by-years .csv file
+    targets::tar_target(
+      pin_agg_csv,
+      PFUWorkflow::release_target(pipeline_releases_folder = pinboard_folder,
+                                  targ = agg_Re_all_St_pfu%>%
+                                    pivot_agg_eta_wide_by_year(pivot_cols = c(IEATools::all_stages$primary,
+                                                                              IEATools::all_stages$final,
+                                                                              IEATools::all_stages$useful)),
+                                  targ_name = "agg_Re_all_St_pfu",
+                                  type = "csv",
+                                  release = release)
+    ),
+
+    # Pin efficiencies as a wide-by-years .csv file
+    targets::tar_target(
+      pin_eta_csv,
+      PFUWorkflow::release_target(pipeline_releases_folder = pinboard_folder,
+                                  targ = eta_Re_all_St_pfu %>%
+                                    pivot_agg_eta_wide_by_year(pivot_cols = c(PFUAggDatabase::efficiency_cols$eta_pf,
+                                                                              PFUAggDatabase::efficiency_cols$eta_fu,
+                                                                              PFUAggDatabase::efficiency_cols$eta_pu)),
+                                  targ_name = "eta_Re_all_St_pfu",
+                                  type = "csv",
+                                  release = release)
+    ),
+
+    # Save the cache for posterity.
+    targets::tar_target(
+      store_cache,
+      PFUWorkflow::stash_cache(pipeline_caches_folder = pipeline_caches_output_folder,
+                               cache_folder = "_targets",
+                               file_prefix = "pfu_agg_workflow_cache_",
+                               dependency = c(agg_Re_all_St_pfu, eta_Re_all_St_pfu))
+    )
   )
 }