+++ R/data.R
@@ -62,3 +62,75 @@
 #' @examples
 #' gross_net_metadata
 "gross_net_metadata"
+
+
+#' PFUAggDatabase efficiency data frame column names
+#'
+#' A string list containing named names of columns in the efficiency data frame of the
+#' PFUAggDatabase.
+#'
+#' @format A string list with `r length(efficiency_cols)` entries.
+#' \describe{
+#' \item{eta_pf}{The name of a column containing primary-to-final efficiencies.}
+#' \item{eta_fu}{The name of a column containing final-to-useful efficiencies.}
+#' \item{eta_pu}{The name of a column containing primary-to-useful efficiencies.}
+#' }
+#'
+#' @examples
+#' efficiency_cols
+"efficiency_cols"
+
+
+#' Output file information
+#'
+#' A string list containing named names of files and tabs for outputs from this pipeline.
+#'
+#' @format A string list with `r length(output_file_info)` entries.
+#' \describe{
+#' \item{agg_eta_filename}{The name of the Excel file containing aggregate energy and exergy values and aggregate efficiencies.}
+#' \item{agg_tabname}{The name of the tab in `agg_eta_filename` containing aggregated energy and exergy values.}
+#' \item{eta_tabname}{The name of the tab in `agg_eta_filename` containing aggregated efficiencies.}
+#' }
+#'
+#' @examples
+#' output_file_info
+"output_file_info"
+
+
+#' Aggregation file tabs
+#'
+#' A string list containing names of tabs for an Excel aggregation file.
+#'
+#' The aggregation file should have several tabs, one for each type of aggregation.
+#'
+#' @format A string list with `r length(aggregation_file_tab_names)` entries.
+#' \describe{
+#' \item{continent_aggregation_tab}{The name of the continent aggregation tab.}
+#' \item{world_aggregation_tab}{The name of the world aggregation tab.}
+#' \item{exiobase_region_aggregation_tab}{The name of the Exiobase region aggregation tab.}
+#' \item{eu.product_aggregation_tab}{The name of the useful product aggregation tab.}
+#' \item{ef.product_aggregation_tab}{The name of the final product aggregation tab.}
+#' \item{destination_aggregation_tab}{The name of the destination aggregation tab.}
+#' }
+#'
+#' @examples
+#' aggregation_file_tab_names
+"aggregation_file_tab_names"
+
+
+#' Aggregation table column names
+#'
+#' A string list containing names of column names in aggregation tables.
+#'
+#' @format A string list with `r length(aggregation_file_cols)` entries.
+#' \describe{
+#' \item{many_colname}{The name of the many column in each tab.}
+#' \item{few_colname}{The name of the few column in each tab.}
+#' }
+#'
+#' @examples
+#' aggregation_file_cols
+"aggregation_file_cols"
+
+
+