+++ _targets.R
@@ -8,31 +8,42 @@ library(targets)
 # Set control parameters for the pipeline.
 
 # Set the countries to be analyzed.
-countries <- c("WMB", "WAB")
+# countries <- c("GBR")
+# countries <- c("WMB", "WAB")
 # countries <- c("USA", "CAN", "GBR", "PRT", "ZAF", "WMB", "WAB")
 # countries <- PFUWorkflow::canonical_countries[1:76] %>% unlist()
-# countries <- "all" # Run all countries
+countries <- "all" # Run all countries
 
-# Set the release of PSUT to be used.
+# Set the years to be analyzed.
+# years <- 1960
+# years <- 2000
+years <- "all"
+
+# Set the release of PSUT to be used for input.
 psut_release <- "20220225T012039Z-c2035"
 
 # world_agg_map needs to be a double-nested list, because the first layer
 # is stripped off by the targets pipeline.
-world_agg_map <- list(list(WLD = c("AMR", "ASA", "EUR", "OCN", "AFR", "BNK")))
+# world_agg_map <- list(list(WLD = c("AMR", "ASA", "EUR", "OCN", "AFR", "BNK")))
 
 # Number of machine cores to use.
 # Set to less than available on your machine.
 # Applies only to tar_make_clustermq().
-# To parallellize the execution of this pipeline, say
+# To parallelize the execution of this pipeline, say
 # targets::tar_make_clustermq(workers = X),
 # where X is the same as the number of cores.
+# num_cores <- 3
 num_cores <- 8
 
 # Set the target to debug.  Set to NULL to turn off debugging.
 # To debug, set appropriate breakpoints and use
 # tar_make(callr_function = NULL).
-debug_target <- "PSUT_Re_all_St_fu"
-# debug_target <- NULL
+# debug_target <- "PSUT_Re_all_St_fu"
+debug_target <- NULL
+
+# Should we do a release of the results?
+release <- FALSE
+
 
 # End user-adjustable parameters.
 
@@ -53,6 +64,7 @@ targets::tar_option_set(
   packages = c(
     "dplyr",
     "IEATools",
+    "parsedate",
     "PFUAggDatabase",
     "pins",
     "tidyr"),
@@ -65,7 +77,9 @@ targets::tar_option_set(
 
 # Pull in the pipeline
 PFUAggDatabase::get_pipeline(countries = countries,
+                             years = years,
                              psut_release = psut_release,
-                             psut_releases_folder = PFUSetup::get_abs_paths()[["workflow_releases_folder"]],
-                             exemplar_table_path = PFUSetup::get_abs_paths()[["exemplar_table_path"]],
-                             world_agg_map = world_agg_map)
+                             aggregation_maps_path = PFUSetup::get_abs_paths()[["aggregation_mapping_path"]],
+                             pipeline_caches_folder = PFUSetup::get_abs_paths()[["pipeline_caches_folder"]],
+                             pipeline_releases_folder = PFUSetup::get_abs_paths()[["pipeline_releases_folder"]],
+                             release = release)