+++ NEWS.md
@@ -4,6 +4,12 @@ output: html_document
 ---
 
 
+# PFUAggDatabase 0.0.1 (2022-04-03)
+
+* Now saving important results to a pinboard.
+* Now saving workflow cache to a .zip file.
+* Added a new `years` variable that allows user-selection of years to be analyzed.
+* Added efficiency calculations with target `write_agg_etas_xlsx`.
 * Added spell-checking to the build process.
 * Added primary, final, and useful energy and exergy targets
   that use all regional aggregations.